'fix' counting after files after the Core::Stream rename
This commit:

- Adds 'Deprecated' prefix for old AK::Streams for commits after
  ae64b68717c02b902721e861dbf51c5d32950193
- Expands Core::Stream regex to work also for SerenityOS/serenity#17173
  (this will also improve counting usages before that patchset where
  `using Core::Stream` was used in FlacLoader and Brotli.)
- Removes some false-positives in AK::DeprecatedStream

ultimately this makes the Core::Stream usage drop by only about 30
after 8464da143987e290fa3d9ad739b491e6adbffee6, mainly because it will
no longer count SeekMode enums, because they are now in the global
namespace.

probably it would be better to make separate sets of regexes after
such things, but it's seems to work so i'll just leave it as is.